+++ packages/algob/src/lib/compile.ts
@@ -1,5 +1,5 @@
 import { getPathFromDirRecursive } from "@algo-builder/runtime";
-import { BuilderError, ERRORS, parseAlgorandError } from "@algo-builder/web";
+import { BuilderError, ERRORS, parseAlgorandError, types } from "@algo-builder/web";
 import type { Algodv2, modelsv2 } from "algosdk";
 import { spawnSync, SpawnSyncReturns } from "child_process";
 import * as fs from 'fs';
@@ -81,7 +81,7 @@ export class CompileOp {
       const p = path.join(CACHE_DIR, filename + ".yaml");
       return YAML.parse(await fs.promises.readFile(p, 'utf8')) as ASCCache;
     } catch (e) {
-      if (e?.errno === -2) { return undefined; } // handling a not existing file
+      if (types.isFileError(e) && e?.errno === -2) { return undefined; } // handling a not existing file
       throw e;
     }
   }
@@ -103,7 +103,8 @@ export class CompileOp {
         base64ToBytes: new Uint8Array(Buffer.from(co.result, "base64"))
       };
     } catch (e) {
-      throw parseAlgorandError(e, { filename: filename });
+      if (types.isRequestError(e)) { throw parseAlgorandError(e, { filename: filename }); }
+      throw e;
     }
   }
 
@@ -206,7 +207,7 @@ export class PyCompileOp {
       const p = path.join(CACHE_DIR, filename + ".yaml");
       return YAML.parse(await fs.promises.readFile(p, 'utf8')) as PyASCCache;
     } catch (e) {
-      if (e?.errno === -2) { return undefined; } // handling a not existing file
+      if (types.isFileError(e) && e?.errno === -2) { return undefined; } // handling a not existing file
       throw e;
     }
   }