+++ packages/runtime/src/interpreter/opcode-list.ts
@@ -9,9 +9,12 @@ import { Keccak } from 'sha3';
 import { RUNTIME_ERRORS } from "../errors/errors-list";
 import { RuntimeError } from "../errors/runtime-errors";
 import { compareArray } from "../lib/compare";
-import { AssetParamMap, GlobalFields, MAX_CONCAT_SIZE, MAX_UINT64, MaxTEALVersion, TxArrFields } from "../lib/constants";
 import {
-  assertLen, assertOnlyDigits, convertToBuffer,
+  AssetParamMap, GlobalFields, MAX_CONCAT_SIZE, MAX_INPUT_BYTE_LEN, MAX_OUTPUT_BYTE_LEN,
+  MAX_UINT64, MaxTEALVersion, TxArrFields
+} from "../lib/constants";
+import {
+  assertLen, assertOnlyDigits, bigEndianBytesToBigInt, bigintToBigEndianBytes, convertToBuffer,
   convertToString, getEncoding, parseBinaryStrToBigInt
 } from "../lib/parsing";
 import { Stack } from "../lib/stack";
@@ -44,7 +47,11 @@ export class Pragma extends Op {
       this.version = Number(args[1]);
       interpreter.tealVersion = this.version;
     } else {
-      throw new RuntimeError(RUNTIME_ERRORS.TEAL.PRAGMA_VERSION_ERROR, { got: args.join(' '), line: line });
+      throw new RuntimeError(RUNTIME_ERRORS.TEAL.PRAGMA_VERSION_ERROR, {
+        expected: 'till #4',
+        got: args.join(' '),
+        line: line
+      });
     }
   }
 
@@ -1394,7 +1401,7 @@ export class Label extends Op {
   execute (stack: TEALStack): void {}
 }
 
-// branch unconditionally to label
+// branch unconditionally to label - Tealv <= 3
 // push to stack [...stack]
 export class Branch extends Op {
   readonly label: string;
@@ -1420,7 +1427,16 @@ export class Branch extends Op {
   }
 }
 
-// branch conditionally if top of stack is zero
+// branch unconditionally to label - TEALv4
+// can also jump backward
+// push to stack [...stack]
+export class Branchv4 extends Branch {
+  execute (stack: TEALStack): void {
+    this.interpreter.jumpToLabel(this.label, this.line);
+  }
+}
+
+// branch conditionally if top of stack is zero - Teal version <= 3
 // push to stack [...stack]
 export class BranchIfZero extends Op {
   readonly label: string;
@@ -1447,6 +1463,20 @@ export class BranchIfZero extends Op {
 
     if (last === 0n) {
       this.interpreter.jumpForward(this.label, this.line);
+    }
+  }
+}
+
+// branch conditionally if top of stack is zero - Tealv4
+// can jump forward also
+// push to stack [...stack]
+export class BranchIfZerov4 extends BranchIfZero {
+  execute (stack: TEALStack): void {
+    this.assertMinStackLen(stack, 1, this.line);
+    const last = this.assertBigInt(stack.pop(), this.line);
+
+    if (last === 0n) {
+      this.interpreter.jumpToLabel(this.label, this.line);
     }
   }
 }
@@ -1478,6 +1508,20 @@ export class BranchIfNotZero extends Op {
 
     if (last !== 0n) {
       this.interpreter.jumpForward(this.label, this.line);
+    }
+  }
+}
+
+// branch conditionally if top of stack is non zero - Tealv4
+// can jump forward as well
+// push to stack [...stack]
+export class BranchIfNotZerov4 extends BranchIfNotZero {
+  execute (stack: TEALStack): void {
+    this.assertMinStackLen(stack, 1, this.line);
+    const last = this.assertBigInt(stack.pop(), this.line);
+
+    if (last !== 0n) {
+      this.interpreter.jumpToLabel(this.label, this.line);
     }
   }
 }
@@ -1600,13 +1644,14 @@ export class AppOptedIn extends Op {
 
   execute (stack: TEALStack): void {
     this.assertMinStackLen(stack, 2, this.line);
-    const appID = this.assertBigInt(stack.pop(), this.line);
-    const accountIndex = this.assertBigInt(stack.pop(), this.line);
+    const appRef = this.assertBigInt(stack.pop(), this.line);
+    const accountRef: StackElem = stack.pop(); // index to tx.accounts[] OR an address directly
 
-    const account = this.interpreter.getAccount(accountIndex, this.line);
+    const account = this.interpreter.getAccount(accountRef, this.line);
     const localState = account.appsLocalState;
 
-    const isOptedIn = localState.get(Number(appID));
+    const appID = this.interpreter.getAppIDByReference(Number(appRef), false, this.line, this);
+    const isOptedIn = localState.get(appID);
     if (isOptedIn) {
       stack.push(1n);
     } else {
@@ -1638,9 +1683,9 @@ export class AppLocalGet extends Op {
   execute (stack: TEALStack): void {
     this.assertMinStackLen(stack, 2, this.line);
     const key = this.assertBytes(stack.pop(), this.line);
-    const accountIndex = this.assertBigInt(stack.pop(), this.line);
+    const accountRef: StackElem = stack.pop();
 
-    const account = this.interpreter.getAccount(accountIndex, this.line);
+    const account = this.interpreter.getAccount(accountRef, this.line);
     const appID = this.interpreter.runtime.ctx.tx.apid ?? 0;
 
     const val = account.getLocalState(appID, key);
@@ -1674,11 +1719,12 @@ export class AppLocalGetEx extends Op {
   execute (stack: TEALStack): void {
     this.assertMinStackLen(stack, 3, this.line);
     const key = this.assertBytes(stack.pop(), this.line);
-    const appID = this.assertBigInt(stack.pop(), this.line);
-    const accountIndex = this.assertBigInt(stack.pop(), this.line);
+    const appRef = this.assertBigInt(stack.pop(), this.line);
+    const accountRef: StackElem = stack.pop();
 
-    const account = this.interpreter.getAccount(accountIndex, this.line);
-    const val = account.getLocalState(Number(appID), key);
+    const appID = this.interpreter.getAppIDByReference(Number(appRef), false, this.line, this);
+    const account = this.interpreter.getAccount(accountRef, this.line);
+    const val = account.getLocalState(appID, key);
     if (val) {
       stack.push(val);
       stack.push(1n);
@@ -1747,18 +1793,12 @@ export class AppGlobalGetEx extends Op {
   execute (stack: TEALStack): void {
     this.assertMinStackLen(stack, 2, this.line);
     const key = this.assertBytes(stack.pop(), this.line);
-    let appIndex = this.assertBigInt(stack.pop(), this.line);
+    // appRef could be index to foreign apps array,
+    // or since v4 an application id that appears in Txn.ForeignApps
+    const appRef = this.assertBigInt(stack.pop(), this.line);
 
-    const foreignApps = this.interpreter.runtime.ctx.tx.apfa;
-    let appID;
-    if (appIndex === 0n) {
-      appID = this.interpreter.runtime.ctx.tx.apid; // zero index means current app
-    } else {
-      this.checkIndexBound(Number(--appIndex), foreignApps as number[], this.line);
-      appID = foreignApps ? foreignApps[Number(appIndex)] : undefined;
-    }
-
-    const val = this.interpreter.getGlobalState(appID as number, key, this.line);
+    const appID = this.interpreter.getAppIDByReference(Number(appRef), true, this.line, this);
+    const val = this.interpreter.getGlobalState(appID, key, this.line);
     if (val) {
       stack.push(val);
       stack.push(1n);
@@ -1793,9 +1833,9 @@ export class AppLocalPut extends Op {
     this.assertMinStackLen(stack, 3, this.line);
     const value = stack.pop();
     const key = this.assertBytes(stack.pop(), this.line);
-    const accountIndex = this.assertBigInt(stack.pop(), this.line);
+    const accountRef: StackElem = stack.pop();
 
-    const account = this.interpreter.getAccount(accountIndex, this.line);
+    const account = this.interpreter.getAccount(accountRef, this.line);
     const appID = this.interpreter.runtime.ctx.tx.apid ?? 0;
 
     // get updated local state for account
@@ -1857,10 +1897,10 @@ export class AppLocalDel extends Op {
   execute (stack: TEALStack): void {
     this.assertMinStackLen(stack, 1, this.line);
     const key = this.assertBytes(stack.pop(), this.line);
-    const accountIndex = this.assertBigInt(stack.pop(), this.line);
+    const accountRef: StackElem = stack.pop();
 
     const appID = this.interpreter.runtime.ctx.tx.apid ?? 0;
-    const account = this.interpreter.getAccount(accountIndex, this.line);
+    const account = this.interpreter.getAccount(accountRef, this.line);
 
     const localState = account.appsLocalState.get(appID);
     if (localState) {
@@ -1931,8 +1971,8 @@ export class Balance extends Op {
 
   execute (stack: TEALStack): void {
     this.assertMinStackLen(stack, 1, this.line);
-    const accountIndex = this.assertBigInt(stack.pop(), this.line);
-    const acc = this.interpreter.getAccount(accountIndex, this.line);
+    const accountRef: StackElem = stack.pop();
+    const acc = this.interpreter.getAccount(accountRef, this.line);
 
     stack.push(BigInt(acc.balance()));
   }
@@ -1965,11 +2005,12 @@ export class GetAssetHolding extends Op {
 
   execute (stack: TEALStack): void {
     this.assertMinStackLen(stack, 2, this.line);
-    const assetId = this.assertBigInt(stack.pop(), this.line);
-    const accountIndex = this.assertBigInt(stack.pop(), this.line);
+    const assetRef = this.assertBigInt(stack.pop(), this.line);
+    const accountRef: StackElem = stack.pop();
 
-    const account = this.interpreter.getAccount(accountIndex, this.line);
-    const assetInfo = account.assets.get(Number(assetId));
+    const account = this.interpreter.getAccount(accountRef, this.line);
+    const assetID = this.interpreter.getAssetIDByReference(Number(assetRef), false, this.line, this);
+    const assetInfo = account.assets.get(assetID);
     if (assetInfo === undefined) {
       stack.push(0n);
       stack.push(0n);
@@ -2023,18 +2064,9 @@ export class GetAssetDef extends Op {
 
   execute (stack: TEALStack): void {
     this.assertMinStackLen(stack, 1, this.line);
-    const foreignAssetsIdx = this.assertBigInt(stack.pop(), this.line);
-    this.checkIndexBound(
-      Number(foreignAssetsIdx),
-      this.interpreter.runtime.ctx.tx.apas as number[], this.line);
-
-    let assetId;
-    if (this.interpreter.runtime.ctx.tx.apas) {
-      assetId = this.interpreter.runtime.ctx.tx.apas[Number(foreignAssetsIdx)];
-    } else {
-      throw new Error("foreign asset id not found");
-    }
-    const AssetDefinition = this.interpreter.getAssetDef(assetId);
+    const assetRef = this.assertBigInt(stack.pop(), this.line);
+    const assetID = this.interpreter.getAssetIDByReference(Number(assetRef), true, this.line, this);
+    const AssetDefinition = this.interpreter.getAssetDef(assetID);
     let def: string;
 
     if (AssetDefinition === undefined) {
@@ -2585,9 +2617,175 @@ export class MinBalance extends Op {
 
   execute (stack: TEALStack): void {
     this.assertMinStackLen(stack, 1, this.line);
-    const accountIndex = this.assertBigInt(stack.pop(), this.line);
-    const acc = this.interpreter.getAccount(accountIndex, this.line);
+    const accountRef: StackElem = stack.pop();
+    const acc = this.interpreter.getAccount(accountRef, this.line);
 
     stack.push(BigInt(acc.minBalance));
+  }
+}
+
+/** TEALv4 Ops **/
+
+// push Ith scratch space index of the Tth transaction in the current group
+// push to stack [...stack, bigint/bytes]
+// Pops nothing
+// Args expected: [{uint8 transaction group index}(T),
+// {uint8 position in scratch space to load from}(I)]
+export class Gload extends Op {
+  readonly scratchIndex: number;
+  txIndex: number;
+  readonly interpreter: Interpreter;
+  readonly line: number;
+
+  /**
+   * Stores scratch space index and transaction index number according to arguments passed.
+   * @param args Expected arguments: [index number]
+   * @param line line number in TEAL file
+   * @param interpreter interpreter object
+   */
+  constructor (args: string[], line: number, interpreter: Interpreter) {
+    super();
+    this.line = line;
+    assertLen(args.length, 2, this.line);
+    assertOnlyDigits(args[0], this.line);
+    assertOnlyDigits(args[1], this.line);
+
+    this.txIndex = Number(args[0]);
+    this.scratchIndex = Number(args[1]);
+    this.interpreter = interpreter;
+  }
+
+  execute (stack: TEALStack): void {
+    const scratch = this.interpreter.runtime.ctx.sharedScratchSpace.get(this.txIndex);
+    if (scratch === undefined) {
+      throw new RuntimeError(
+        RUNTIME_ERRORS.TEAL.SCRATCH_EXIST_ERROR,
+        { index: this.txIndex, line: this.line }
+      );
+    }
+    this.checkIndexBound(this.scratchIndex, scratch, this.line);
+    stack.push(scratch[this.scratchIndex]);
+  }
+}
+
+// push Ith scratch space index of the Tth transaction in the current group
+// push to stack [...stack, bigint/bytes]
+// Pops uint64(T)
+// Args expected: [{uint8 position in scratch space to load from}(I)]
+export class Gloads extends Gload {
+  /**
+   * Stores scratch space index number according to argument passed.
+   * @param args Expected arguments: [index number]
+   * @param line line number in TEAL file
+   * @param interpreter interpreter object
+   */
+  constructor (args: string[], line: number, interpreter: Interpreter) {
+    // "11" is mock value, will be updated when poping from stack in execute
+    super(["11", ...args], line, interpreter);
+  }
+
+  execute (stack: TEALStack): void {
+    this.assertMinStackLen(stack, 1, this.line);
+    this.txIndex = Number(this.assertBigInt(stack.pop(), this.line));
+    super.execute(stack);
+  }
+}
+
+/**
+ * Provide subroutine functionality. When callsub is called, the current location in
+ * the program is saved and immediately jumps to the label passed to the opcode.
+ * Pops: None
+ * Pushes: None
+ * The call stack is separate from the data stack. Only callsub and retsub manipulate it.
+ * Pops: None
+ * Pushes: Pushes current instruction index in call stack
+ */
+export class Callsub extends Op {
+  readonly interpreter: Interpreter;
+  readonly label: string;
+  readonly line: number;
+
+  /**
+   * Sets `label` according to arguments passed.
+   * @param args Expected arguments: [label of branch]
+   * @param line line number in TEAL file
+   * @param interpreter interpreter object
+   */
+  constructor (args: string[], line: number, interpreter: Interpreter) {
+    super();
+    assertLen(args.length, 1, line);
+    this.label = args[0];
+    this.interpreter = interpreter;
+    this.line = line;
+  }
+
+  execute (stack: TEALStack): void {
+    // the current location in the program is saved
+    this.interpreter.callStack.push(this.interpreter.instructionIndex);
+    // immediately jumps to the label passed to the opcode.
+    this.interpreter.jumpToLabel(this.label, this.line);
+  }
+}
+
+/**
+ * When the retsub opcode is called, the AVM will resume
+ * execution at the previous saved point.
+ * Pops: None
+ * Pushes: None
+ * The call stack is separate from the data stack. Only callsub and retsub manipulate it.
+ * Pops: index from call stack
+ * Pushes: None
+ */
+export class Retsub extends Op {
+  readonly interpreter: Interpreter;
+  readonly line: number;
+
+  /**
+   * @param args Expected arguments: []
+   * @param line line number in TEAL file
+   * @param interpreter interpreter object
+   */
+  constructor (args: string[], line: number, interpreter: Interpreter) {
+    super();
+    assertLen(args.length, 0, line);
+    this.interpreter = interpreter;
+    this.line = line;
+  }
+
+  execute (stack: TEALStack): void {
+    // get current location from saved point
+    // jump to saved instruction opcode
+    if (this.interpreter.callStack.length() === 0) {
+      throw new RuntimeError(RUNTIME_ERRORS.TEAL.CALL_STACK_EMPTY, { line: this.line });
+    }
+    this.interpreter.instructionIndex = this.interpreter.callStack.pop();
+  }
+}
+
+// A plus B, where A and B are byte-arrays interpreted as big-endian unsigned integers
+// panics on overflow (result > max_uint1024 i.e 128 byte num)
+// Pops: ... stack, {[]byte A}, {[]byte B}
+// push to stack [...stack, []byte]
+export class ByteAdd extends Op {
+  readonly line: number;
+  /**
+   * Asserts 0 arguments are passed.
+   * @param args Expected arguments: [] // none
+   * @param line line number in TEAL file
+   */
+  constructor (args: string[], line: number) {
+    super();
+    this.line = line;
+    assertLen(args.length, 0, line);
+  };
+
+  execute (stack: TEALStack): void {
+    this.assertMinStackLen(stack, 2, this.line);
+    const byteB = this.assertBytes(stack.pop(), this.line, MAX_INPUT_BYTE_LEN);
+    const byteA = this.assertBytes(stack.pop(), this.line, MAX_INPUT_BYTE_LEN);
+
+    const result = bigEndianBytesToBigInt(byteA) + bigEndianBytesToBigInt(byteB);
+    const resultAsBytes = bigintToBigEndianBytes(result);
+    stack.push(this.assertBytes(resultAsBytes, this.line, MAX_OUTPUT_BYTE_LEN));
   }
 }