+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # CHANGELOG
 
+## Unreleased
+
+### Improvements
++ Added shared space between contracts
++ Added tealv4 opcodes (`gload` and `gloads`) 
++ Added Tealv4 opcodes (`callsub` and `retsub`)
++ Added loop support in runtime
+* TEALv4 support in `@algo-builder/runtime`:
+    * Added shared space between contracts (opcodes `gload` and `gloads`)
+    * Dynamic Opcode Cost Evaluation
+    * Transaction Array changes
+        a) array length assertions for `tx.ForeignAssets`, `tx.Accounts`, `tx.ForeignApps`,
+        b) User can pass id/offset for app/asset in for array references. For `tx.Accounts` you can pass address directly in teal code.
+
+
+### Bug Fixes
+* Fixed `yarn add @algo-builder/web` (was failing because of missing dependency `zod` in packages/web).
+* Fix metadatahash type
+
+### API Breaking
+* Migrate from `LogicSig` to `LogicSigAccount`(Note: Loading lsig from file uses `LogicSig`, because `goal` stores it in LogicSig type format)
+
+### API breaking
+* Rename `CallNoOpSSC` to `CallApp`.
+
 ## v1.2.0 2021-08-09
 
 ### Improvements